# application/services/custom_services/crudticket/src/models/entitymodels/students.ts
export interface students 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   email: { type: String }
}